Add support for "-m" for cache-miss (Fixes #92)
Add support for "-m" / "--cache-miss" flag to prepend fqdn in request
with a random 8-bytes string to enforce a cache miss.

This will help users to measure delay the whole recursion chain, by
enforcing DNS resolver to contact "resolve" the name every time.
As a consequence, the response may be empty or contain a wildcard
response.